+++ src/views/demo/draggable-card.tsx
@@ -5,30 +5,40 @@ export default defineComponent({
   name: 'DraggableCardDemo',
   setup() {
     const card3 = ref(false)
+    const domRef = useTemplateRef<HTMLElement>('domRef')
 
     return {
       card3,
+      domRef,
     }
   },
   render() {
-    const { card3 } = this
+    const { card3, domRef } = this
 
     return (
       <div>
+        <div
+          ref="domRef"
+          style={{
+            width: '100%',
+            height: '400px',
+            backgroundColor: 'red',
+          }}
+        ></div>
         <RDraggableCard animation title="Body">
           我被限制在 body 中。
         </RDraggableCard>
         {card3 ? (
           <RDraggableCard
             animation
-            title="Layout Container"
-            restrictionElement=".r-layout-full__viewer-content"
+            title="Custom Container"
+            restrictionElement={domRef}
             closable
             onClose={() => (this.card3 = false)}
           >
             {{
               default: () =>
-                '我被限制在 Layout Container 中。并且我支持 NCard 的所有配置与插槽。',
+                '我被限制在红色区域中。并且我支持 NCard 的所有配置与插槽。',
               'header-extra': () => '其实我就是 NCard 封装的',
               footer: () => '我支持 footer 插槽',
               action: () => '我支持 action 插槽',